feat(BPAAS-2199): consider number of blocks when rejecting pending txs

+++ config/config.go
@@ -107,7 +107,7 @@ type MetamorphConfig struct {
 type RejectPendingSeenConfig struct {
 	Enabled          bool          `mapstructure:"enabled"`
 	LastRequestedAgo time.Duration `mapstructure:"lastRequestedAgo"`
-	BlocksSince      int           `mapstructure:"blocksSince"`
+	BlocksSince      uint64        `mapstructure:"blocksSince"`
 }
 
 type ReAnnounceSeenConfig struct {
